pyproject.toml
[project]
name = "voter-hak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-jazzmin>=3.0.4",
]
